requirements.txt
aniso8601==9.0.1
click==8.1.7
Flask==2.2.5
Flask-RESTful==0.3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==0.21.1
pytz==2023.3.post1
six==1.16.0
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
aniso8601==9.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
environs==9.5.0
Flask==2.2.5
Flask-RESTful==0.3.10
idna==3.6
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
minio==7.2.0
numpy==1.21.6
packaging==23.2
pandas==1.3.5
protobuf==4.24.4
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.19.0
pymilvus==2.3.3
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
ujson==5.7.0
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
aniso8601==9.0.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
environs==9.5.0
Flask==2.2.5
Flask-RESTful==0.3.10
idna==3.6
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
minio==7.2.0
numpy==1.21.6
packaging==23.2
pandas==1.3.5
protobuf==4.24.4
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.19.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
ujson==5.7.0
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
